+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.28.0] - 2020-07-02
+
 ### Addded
 
 - Copy and link files from the ExtData folder when importing relANNIS.